Remove unused codes related to MonsterCollection
