+++ .Lib9c.Tests/Action/ActionEvaluationTest.cs
@@ -53,7 +53,6 @@ namespace Lib9c.Tests.Action
         [InlineData(typeof(AddRedeemCode))]
         [InlineData(typeof(Buy))]
         [InlineData(typeof(ChargeActionPoint))]
-        [InlineData(typeof(ClaimMonsterCollectionReward))]
         [InlineData(typeof(CombinationConsumable))]
         [InlineData(typeof(CombinationEquipment))]
         [InlineData(typeof(CreatePendingActivation))]
@@ -64,7 +63,6 @@ namespace Lib9c.Tests.Action
         [InlineData(typeof(MigrationAvatarState))]
         [InlineData(typeof(MigrationLegacyShop))]
         [InlineData(typeof(MimisbrunnrBattle))]
-        [InlineData(typeof(MonsterCollect))]
         [InlineData(typeof(PatchTableSheet))]
         [InlineData(typeof(RankingBattle))]
         [InlineData(typeof(RapidCombination))]
@@ -180,7 +178,6 @@ namespace Lib9c.Tests.Action
                     },
                 },
                 ChargeActionPoint _ => new ChargeActionPoint(),
-                ClaimMonsterCollectionReward _ => new ClaimMonsterCollectionReward(),
                 CombinationConsumable _ => new CombinationConsumable(),
                 CombinationEquipment _ => new CombinationEquipment(),
                 CreatePendingActivation _ => new CreatePendingActivation(
@@ -214,7 +211,6 @@ namespace Lib9c.Tests.Action
                     PlayCount = 0,
                     AvatarAddress = default,
                 },
-                MonsterCollect _ => new MonsterCollect(),
                 PatchTableSheet _ => new PatchTableSheet
                 {
                     TableCsv = "table",
