+++ Lib9c/Action/MigrateMonsterCollection.cs
@@ -1,10 +1,14 @@
 using System;
+using System.Collections.Generic;
 using Bencodex.Types;
 using Lib9c.Abstractions;
 using Libplanet.Action;
 using Libplanet.Action.State;
 using Libplanet.Crypto;
+using Nekoyume.Model.Item;
+using Nekoyume.Model.Mail;
 using Nekoyume.Model.State;
+using Nekoyume.TableData;
 using Serilog;
 using static Lib9c.SerializeKeys;
 
@@ -55,7 +59,7 @@ namespace Nekoyume.Action
 
             try
             {
-                states = ClaimMonsterCollectionReward.Claim(context, AvatarAddress, addressesHex);
+                states = ClaimMonsterCollectionReward(context, AvatarAddress, addressesHex);
             }
             catch (Exception e)
             {
@@ -92,6 +96,68 @@ namespace Nekoyume.Action
                     monsterCollectionState.address,
                     migratedStakeStateAddress,
                     states.GetBalance(monsterCollectionState.address, currency));
+        }
+
+        private static IAccount ClaimMonsterCollectionReward(IActionContext context, Address avatarAddress, string addressesHex)
+        {
+            const long MonsterCollectionRewardEndBlockIndex = 4_481_909;
+
+            IAccount states = context.PreviousState;
+            Address inventoryAddress = avatarAddress.Derive(LegacyInventoryKey);
+            Address worldInformationAddress = avatarAddress.Derive(LegacyWorldInformationKey);
+            Address questListAddress = avatarAddress.Derive(LegacyQuestListKey);
+            var started = DateTimeOffset.UtcNow;
+            Log.Debug("{AddressesHex}ClaimMonsterCollection exec started", addressesHex);
+
+            if (!states.TryGetAgentAvatarStatesV2(context.Signer, avatarAddress, out AgentState agentState, out AvatarState avatarState, out _))
+            {
+                throw new FailedLoadStateException($"Aborted as the avatar state of the signer failed to load.");
+            }
+
+            Address collectionAddress = MonsterCollectionState.DeriveAddress(context.Signer, agentState.MonsterCollectionRound);
+
+            if (!states.TryGetState(collectionAddress, out Dictionary stateDict))
+            {
+                throw new FailedLoadStateException($"Aborted as the monster collection state failed to load.");
+            }
+
+            var monsterCollectionState = new MonsterCollectionState(stateDict);
+            List<MonsterCollectionRewardSheet.RewardInfo> rewards =
+                monsterCollectionState.CalculateRewards(
+                    states.GetSheet<MonsterCollectionRewardSheet>(),
+                    Math.Min(MonsterCollectionRewardEndBlockIndex, context.BlockIndex)
+                );
+
+            if (rewards.Count == 0)
+            {
+                throw new RequiredBlockIndexException($"{collectionAddress} is not available yet");
+            }
+
+            var random = context.GetRandom();
+            Guid id = random.GenerateRandomGuid();
+            var result = new MonsterCollectionResult(id, avatarAddress, rewards);
+            var mail = new MonsterCollectionMail(result, context.BlockIndex, id, context.BlockIndex);
+            avatarState.Update(mail);
+
+            ItemSheet itemSheet = states.GetItemSheet();
+            foreach (MonsterCollectionRewardSheet.RewardInfo rewardInfo in rewards)
+            {
+                ItemSheet.Row row = itemSheet[rewardInfo.ItemId];
+                ItemBase item = row is MaterialItemSheet.Row materialRow
+                    ? ItemFactory.CreateTradableMaterial(materialRow)
+                    : ItemFactory.CreateItem(row, random);
+                avatarState.inventory.AddItem(item, rewardInfo.Quantity);
+            }
+            monsterCollectionState.Claim(context.BlockIndex);
+
+            var ended = DateTimeOffset.UtcNow;
+            Log.Debug("{AddressesHex}ClaimMonsterCollection Total Executed Time: {Elapsed}", addressesHex, ended - started);
+            return states
+                .SetState(avatarAddress, avatarState.SerializeV2())
+                .SetState(inventoryAddress, avatarState.inventory.Serialize())
+                .SetState(worldInformationAddress, avatarState.worldInformation.Serialize())
+                .SetState(questListAddress, avatarState.questList.Serialize())
+                .SetState(collectionAddress, monsterCollectionState.Serialize());
         }
     }
 }
